backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloud-sql-python-connector==1.13.0
cryptography==43.0.3
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
google-auth==2.36.0
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
multidict==6.1.0
mysqlclient==2.2.6
propcache==0.2.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
yarl==1.17.1
